Update atividade01.c
Changed float variables to double for improved precision.

+++ aula-01/atividade01.c
@@ -3,8 +3,8 @@
 
 int main() {
 
-    float nota1[30], nota2[30], media[30];
-    float soma = 0, mediaTurma;
+    double nota1[30], nota2[30], media[30];
+    double soma = 0, mediaTurma;
     int i, resultado;
 
     for (i = 0; i < 30; i++) {
@@ -15,7 +15,7 @@ int main() {
         while (1) {
 
             printf("Digite a primeira nota (0 a 10): ");
-            resultado = scanf("%f", &nota1[i]);
+            resultado = scanf("%lf", &nota1[i]);
 
             if (resultado != 1) {
                 printf("Nota invalida, por favor digite o numero novamente.\n");
@@ -35,7 +35,7 @@ int main() {
         while (1) {
 
             printf("Digite a segunda nota (0 a 10): ");
-            resultado = scanf("%f", &nota2[i]);
+            resultado = scanf("%lf", &nota2[i]);
 
             if (resultado != 1) {
                 printf("Nota invalida, por favor digite o numero novamente.\n");
@@ -57,7 +57,7 @@ int main() {
 
     mediaTurma = soma / 30;
 
-    printf("\nMedia da turma: %.2f\n", mediaTurma);
+    printf("\nMedia da turma: %.2lf\n", mediaTurma);
 
     printf("\nAlunos acima da media da turma:\n");
 
@@ -66,9 +66,9 @@ int main() {
         if (media[i] > mediaTurma) {
 
             printf("\nAluno %d\n", i + 1);
-            printf("Nota 1: %.2f\n", nota1[i]);
-            printf("Nota 2: %.2f\n", nota2[i]);
-            printf("Media: %.2f\n", media[i]);
+            printf("Nota 1: %.2lf\n", nota1[i]);
+            printf("Nota 2: %.2lf\n", nota2[i]);
+            printf("Media: %.2lf\n", media[i]);
         }
     }
